fix: item_count Int crash, dead code, formatDuration duplication
- server/shows.js: cast COUNT()::integer so pg driver sends a number
  instead of string — fixes kotlinx deserialization crash on /api/shows
- ApiClient.kt: add coerceInputValues=true as defense against future
  type mismatches from numeric DB columns
- Navigation.kt: remove unused `val show = state.showDetail`
- Extract formatDuration to ui/util/FormatDuration.kt and import it
  in PlayerScreen + ShowDetailScreen — removes identical private copies

+++ server/src/routes/shows.js
@@ -8,7 +8,7 @@ export default async function showRoutes(fastify) {
 
     let q = `
       SELECT s.*,
-        COUNT(m.id) as item_count,
+        COUNT(m.id)::integer as item_count,
         MAX(m.scanned_at) as last_added
       FROM shows s
       LEFT JOIN media_items m ON m.show_id = s.id
